Better organization for AsyncScheduler code
