--- Sources/Player/Common/World/AsyncSchedulerFactoryProvider.swift
@@ -1,13 +1,0 @@
-import Foundation
-
-// MARK: - AsyncSchedulerFactoryProvider.swift
-
-struct AsyncSchedulerFactoryProvider {
-	var newFactory: () -> AsyncSchedulerFactory
-}
-
-extension AsyncSchedulerFactoryProvider {
-	public static let live = AsyncSchedulerFactoryProvider(
-		newFactory: { TaskAsyncSchedulerFactory() }
-	)
-}
