+++ Sources/Player/Common/World/AsyncSchedulerFactoryProvider/AsyncSchedulerFactoryProvider+Live.swift
@@ -1,0 +1,5 @@
+extension AsyncSchedulerFactoryProvider {
+	public static let live = AsyncSchedulerFactoryProvider(
+		newFactory: { TaskAsyncSchedulerFactory() }
+	)
+}
