+++ Sources/Player/Common/World/AsyncSchedulerFactoryProvider/AsyncSchedulerFactoryProvider.swift
@@ -1,0 +1,7 @@
+import Foundation
+
+/// Provider of AsyncSchedulerFactory used to dispatch Tasks in a controlled way.
+
+struct AsyncSchedulerFactoryProvider {
+	var newFactory: () -> AsyncSchedulerFactory
+}
