--- Sources/Player/PlaybackEngine/Internal/Events/AsyncScheduler.swift
@@ -1,23 +1,0 @@
-protocol AsyncScheduler {
-  static func schedule(code: @escaping () async -> Void) -> AsyncScheduler
-}
-
-protocol AsyncSchedulerFactory {
-  @discardableResult
-  func create(code: @escaping () async -> Void) -> AsyncScheduler
-}
-
-extension Task: AsyncScheduler {
-  static func schedule(code: @escaping () async -> Void) -> AsyncScheduler {
-	SafeTask {
-	  await code()
-	}
-  }
-}
-
-final class TaskAsyncSchedulerFactory: AsyncSchedulerFactory {
-  @discardableResult
-  func create(code: @escaping () async -> Void) -> AsyncScheduler {
-	Task<Void, Never>.schedule(code: code)
-  }
-}
