+++ Sources/Player/PlaybackEngine/Internal/Events/AsyncScheduler/AsyncScheduler+Task.swift
@@ -1,0 +1,7 @@
+extension Task: AsyncScheduler {
+  static func schedule(code: @escaping () async -> Void) -> AsyncScheduler {
+	Task<Void, Never> {
+	  await code()
+	}
+  }
+}
