+++ Sources/Player/PlaybackEngine/Internal/Events/AsyncScheduler/AsyncScheduler.swift
@@ -1,0 +1,3 @@
+protocol AsyncScheduler {
+  static func schedule(code: @escaping () async -> Void) -> AsyncScheduler
+}
