+++ Sources/Player/PlaybackEngine/Internal/Events/AsyncScheduler/AsyncSchedulerFactory.swift
@@ -1,0 +1,4 @@
+protocol AsyncSchedulerFactory {
+  @discardableResult
+  func create(code: @escaping () async -> Void) -> AsyncScheduler
+}
