+++ Sources/Player/PlaybackEngine/Internal/Events/AsyncScheduler/TaskAsyncSchedulerFactory.swift
@@ -1,0 +1,8 @@
+import Foundation
+
+final class TaskAsyncSchedulerFactory: AsyncSchedulerFactory {
+  @discardableResult
+  func create(code: @escaping () async -> Void) -> AsyncScheduler {
+	Task<Void, Never>.schedule(code: code)
+  }
+}
